refactor(construction): split armorCalculations into focused modules
